Moved ImageCommentsErrorViewModel, ImageCommentsLoadingViewModel, ImageCommentsPresenter, PresentableImageComment to their own files for production.
