+++ EssentialFeed/EssentialFeed/Image Comments/Presenter/ImageCommentsErrorViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ImageCommentsErrorViewModel.swift
+//  EssentialFeed
+//
+//  Created by Rakesh Ramamurthy on 01/12/20.
+//  Copyright © 2020 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public struct ImageCommentsErrorViewModel {
+	public let errorMessage: String?
+}
