+++ EssentialFeed/EssentialFeed/Image Comments/Presenter/ImageCommentsLoadingViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ImageCommentsLoadingViewModel.swift
+//  EssentialFeed
+//
+//  Created by Rakesh Ramamurthy on 01/12/20.
+//  Copyright © 2020 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public struct ImageCommentsLoadingViewModel {
+	public let isLoading: Bool
+}
