+++ EssentialFeed/EssentialFeed/Image Comments/Presenter/el.lproj/ImageComments.strings
@@ -5,6 +5,5 @@
   Created by Rakesh Ramamurthy on 01/12/20.
   Copyright © 2020 Essential Developer. All rights reserved.
 */
-
 "IMAGE_COMMENTS_VIEW_TITLE" = "Σχόλια";
 "IMAGE_COMMENTS_VIEW_CONNECTION_ERROR" = "Δεν ήταν δυνατή η σύνδεση στον διακομιστή";
